fix: Mocha/Chai requires done param to test asynch func
When testing an asynch function using Mocha/Chai,
The test function (the one passed into the describe function)
Must take in a "done" parameter,
And the anon function that we pass into the asynch func that we are testing
Must call done().

This way the Mocha/Chai test is run from inside the asynch's callback function,
Ensureing an accurate test result.

(Otherwise, the test is run before the asynch part ever executes its callback
- ie before we get a reply back from our API call, or whatever the asynch thing is.)